refactor into tool functions to make it faster
Prior to that commit, the script had to pull from the GitHub API
each time an information was required. This is highly inefficient
as all the information is already contained in the first
`gh api notifications --jq '.'`! But it was discard each time...
This commit refactors the script into tool sub-functions and, most
importantly, stores the results of the first `gh api notifications`
inside a temporary file on the filesystem, making the following
pulls much faster!